feat(layout): move global initialization to app
